refactoring and init the db obj in index.ts
