+++ src/index.ts
@@ -1,1 +1,6 @@
-export { imongo } from './imongo';
+import { imongo } from './core';
+
+const db = new imongo();
+
+export { z } from 'zod';
+export { db };
